[SPARK-14798][SQL] Move native command and script transformation parsing into SparkSqlAstBuilder
